# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API-сервер для управления заявками и продуктами в системе DietComposition.",
        "title": "API для системы DietComposition",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8090",
    "basePath": "/api",
    "paths": {
...
        "/diet_composition": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает отфильтрованный список всех сформированных заявок (кроме черновиков и удаленных).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "diet_composition"
                ],
                "summary": "Получить список заявок (авторизованный пользователь)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Фильтр по статусу заявки",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по дате 'от' (формат YYYY-MM-DD)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по дате 'до' (формат YYYY-MM-DD)",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ds.DietCompositionDTO"
                            }
                        }
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/diet_composition/cart": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает ID черновика текущего пользователя и количество продуктов в нем.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "diet_composition"
                ],
                "summary": "Получить информацию для иконки корзины (авторизованный пользователь)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.CartBadgeDTO"
                        }
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/diet_composition/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает полную информацию о заявке, включая привязанные продукты.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "diet_composition"
                ],
                "summary": "Получить одну заявку по ID (авторизованный пользователь)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.DietCompositionDTO"
                        }
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Заявка не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Позволяет пользователю обновить поля своей заявки .",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "diet_composition"
                ],
                "summary": "Обновить данные заявки (авторизованный пользователь)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления",
                        "name": "updateData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.DietCompositionUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Логически удаляет заявку, переводя ее в статус \"удалена\".",
                "tags": [
                    "diet_composition"
                ],
                "summary": "Удалить заявку (авторизованный пользователь)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/diet_composition/{id}/form": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Переводит заявку из статуса \"черновик\" в \"сформирована\".",
                "tags": [
                    "diet_composition"
                ],
                "summary": "Сформировать заявку (авторизованный пользователь)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки (черновика)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Не все поля заполнены",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/diet_composition/{id}/products/{product_id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Изменяет дополнительное описание для конкретного продукта в рамках одной заявки.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "m-m"
                ],
                "summary": "Обновить описание продукта в заявке (авторизованный пользователь)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID продукта",
                        "name": "product_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новое описание",
                        "name": "updateData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.ProductToDietCompositionUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаляет связь между заявкой и продуктом.",
                "tags": [
                    "m-m"
                ],
                "summary": "Удалить продукт из заявки (авторизованный пользователь)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID продукта",
                        "name": "product_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/diet_composition/{id}/resolve": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Модератор завершает (с расчетом) или отклоняет заявку.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "diet_composition"
                ],
                "summary": "Завершить или отклонить заявку (только модератор)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Действие: 'complete' или 'reject'",
                        "name": "action",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.DietCompositionResolveRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ds.CartBadgeDTO": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "diet_composition_id": {
                    "type": "integer"
                }
            }
        },
        "ds.DietCompositionDTO": {
            "type": "object",
            "properties": {
                "PGP": {
                    "type": "number"
                },
                "PRP": {
                    "type": "number"
                },
                "c_pol": {
                    "type": "integer"
                },
                "complition_date": {
                    "type": "string"
                },
                "creation_date": {
                    "type": "string"
                },
                "creator_login": {
                    "type": "integer"
                },
                "forming_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "moderator_login": {
                    "type": "integer"
                },
                "n_pol": {
                    "type": "integer"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.ProductInDietCompositionDTO"
                    }
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "ds.DietCompositionResolveRequest": {
            "type": "object",
            "required": [
                "action"
            ],
            "properties": {
                "action": {
                    "description": "\"complete\" | \"reject\"",
                    "type": "string"
                }
            }
        },
        "ds.DietCompositionUpdateRequest": {
            "type": "object",
            "required": [
                "c_pol",
                "n_pol"
            ],
            "properties": {
                "PGP": {
                    "type": "number"
                },
                "PRP": {
                    "type": "number"
                },
                "c_pol": {
                    "type": "integer"
                },
                "n_pol": {
                    "type": "integer"
                }
            }
        },
...
        "ds.ProductInDietCompositionDTO": {
            "type": "object",
            "properties": {
                "c_pol": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "n_pol": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "ratio": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "ds.ProductToDietCompositionUpdateRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}